cleanup, split webpack plugins into their own module
